requirements.txt
boto3
watchdog
python-dotenv
requests
m3u8
yt-dlp
